--- alembic/versions/096d037beb21_now_head.py
@@ -1,114 +1,0 @@
-"""now head
-
-Revision ID: 096d037beb21
-Revises:
-Create Date: 2025-12-15 15:53:03.726316
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-import sqlmodel.sql
-import sqlmodel.sql.sqltypes
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "096d037beb21"
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "datasets",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-        sa.Column("description", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-        schema="timeseries",
-    )
-    op.create_index(op.f("ix_timeseries_datasets_name"), "datasets", ["name"], unique=True, schema="timeseries")
-    op.create_table(
-        "analyses",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("dataset_id", sa.Integer(), nullable=False),
-        sa.Column("detection_method", sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-        sa.Column("description", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column(
-            "status",
-            sa.Enum("pending", "processing", "completed", "error", name="statustype", schema="timeseries"),
-            nullable=True,
-        ),
-        sa.ForeignKeyConstraint(
-            ["dataset_id"],
-            ["timeseries.datasets.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        schema="timeseries",
-    )
-    op.create_index(
-        op.f("ix_timeseries_analyses_dataset_id"), "analyses", ["dataset_id"], unique=False, schema="timeseries"
-    )
-    op.create_table(
-        "datapoints",
-        sa.Column("dataset_id", sa.Integer(), nullable=False),
-        sa.Column("time", sa.DateTime(), nullable=False),
-        sa.Column("value", sa.Float(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["dataset_id"],
-            ["timeseries.datasets.id"],
-        ),
-        sa.PrimaryKeyConstraint("dataset_id", "time"),
-        schema="timeseries",
-    )
-    op.create_table(
-        "anomalies",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("analysis_id", sa.Integer(), nullable=False),
-        sa.Column("start", sa.DateTime(), nullable=False),
-        sa.Column("end", sa.DateTime(), nullable=False),
-        sa.Column("validated", sa.Boolean(), nullable=False),
-        sa.Column("type", sa.Enum("point", "contextual", name="anomalytype", schema="timeseries"), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["analysis_id"],
-            ["timeseries.analyses.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        schema="timeseries",
-    )
-    op.create_index(
-        op.f("ix_timeseries_anomalies_analysis_id"), "anomalies", ["analysis_id"], unique=False, schema="timeseries"
-    )
-    op.create_table(
-        "prediction_results",
-        sa.Column("analysis_id", sa.Integer(), nullable=False),
-        sa.Column("time", sa.DateTime(), nullable=False),
-        sa.Column("value", sa.Float(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["analysis_id"],
-            ["timeseries.analyses.id"],
-        ),
-        sa.PrimaryKeyConstraint("analysis_id", "time"),
-        schema="timeseries",
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("prediction_results", schema="timeseries")
-    op.drop_index(op.f("ix_timeseries_anomalies_analysis_id"), table_name="anomalies", schema="timeseries")
-    op.drop_table("anomalies", schema="timeseries")
-    op.drop_table("datapoints", schema="timeseries")
-    op.drop_index(op.f("ix_timeseries_analyses_dataset_id"), table_name="analyses", schema="timeseries")
-    op.drop_table("analyses", schema="timeseries")
-    op.drop_index(op.f("ix_timeseries_datasets_name"), table_name="datasets", schema="timeseries")
-    op.drop_table("datasets", schema="timeseries")
-    # ### end Alembic commands ###
