+++ deployments/docker-compose.yml
@@ -39,7 +39,7 @@ services:
   forecasting:
     build:
       context: ..
-      dockerfile: projects/forecasting/Containerfile
+      dockerfile: projects/forecasting_api/Containerfile
     env_file: ".env"
     environment:
       DATABASE_URL: postgresql://${DATABASE_USERNAME}:${DATABASE_PASSWORD}@${DATABASE_HOST}:${DATABASE_PORT}/${DATABASE_NAME}
